#347: Update the frontend to use the new backend API for adding/removing a candidate from a BPS
