+++ apps/frontend/src/swr/ballotPaperSections/useUpdateCandidateInBallotPaperSection.ts
@@ -1,19 +1,29 @@
-import type {
-  SelectableBallotPaper,
-  SelectableBallotPaperSection,
-  SelectableCandidate,
-  SelectableElection,
+import {
+  type SelectableBallotPaper,
+  type SelectableBallotPaperSection,
+  selectableBallotPaperSectionObject,
+  type SelectableCandidate,
+  type SelectableElection,
+  type updateableCandidateOperationOptions,
 } from '@repo/votura-validators';
 import { mutate } from 'swr';
 import useSWRMutation, { type SWRMutationResponse } from 'swr/mutation';
 import { apiRoutes } from '../apiRoutes.ts';
-import { deleter } from '../deleter.ts';
+import { putterFactory } from '../putterFactory.ts';
 
-export const useRemoveCandidateFromBallotPaperSection = (
+export const useUpdateCandidateInBallotPaperSection = (
   electionId?: SelectableElection['id'],
   ballotPaperId?: SelectableBallotPaper['id'],
   ballotPaperSectionId?: SelectableBallotPaperSection['id'],
-): SWRMutationResponse<null, Error, string, { candidateId: SelectableCandidate['id'] }> => {
+): SWRMutationResponse<
+  SelectableBallotPaperSection,
+  Error,
+  string,
+  {
+    candidateId: SelectableCandidate['id'];
+    operation: keyof typeof updateableCandidateOperationOptions;
+  }
+> => {
   const shouldFetch =
     electionId !== undefined && ballotPaperId !== undefined && ballotPaperSectionId !== undefined;
 
@@ -25,7 +35,7 @@ export const useRemoveCandidateFromBallotPaperSection = (
           ballotPaperSectionId,
         )
       : null,
-    deleter,
+    putterFactory(selectableBallotPaperSectionObject),
     {
       onSuccess: () => {
         if (electionId !== undefined && ballotPaperId !== undefined) {
